+++ uui-components/src/overlays/DropdownContainer.tsx
@@ -7,6 +7,7 @@ import {
 import { VPanel } from '../layout/flexItems/VPanel';
 import PopoverArrow from './PopoverArrow';
 import { ReactFocusLockProps } from 'react-focus-lock';
+import { useCallback, useState } from 'react';
 
 export interface DropdownContainerProps
     extends IHasCX,
@@ -56,11 +57,18 @@ export const DropdownContainer = React.forwardRef((props: DropdownContainerProps
         returnFocus = true,
         persistentFocus = true,
     } = props;
+    const [shadowRootHost, setShadowRootHost] = useState<ShadowRoot | null>(null);
+
+    const saveShadowRootHost = useCallback((node: HTMLElement | null) => {
+        if (!node) return;
+        const root = node.getRootNode();
+        setShadowRootHost(root instanceof ShadowRoot ? root : null);
+    }, []);
 
     function renderDropdownContainer() {
         return (
             <VPanel
-                forwardedRef={ !focusLock ? ref as React.ForwardedRef<HTMLDivElement> : undefined }
+                forwardedRef={ !focusLock ? (ref as React.ForwardedRef<HTMLDivElement>) : saveShadowRootHost }
                 cx={ cx(uuiElement.dropdownBody, uuiMarkers.lockFocus, props.cx) }
                 style={ {
                     ...props.style,
@@ -91,8 +99,9 @@ export const DropdownContainer = React.forwardRef((props: DropdownContainerProps
                 persistentFocus={ persistentFocus }
                 lockProps={ { ...({ onKeyDown: props?.onKeyDown }), ...props.lockProps } }
                 shards={ props.shards }
-                autoFocus={ props.autoFocus || true }
+                autoFocus={ props.autoFocus ?? true }
                 as={ props.as }
+                shadowRootHost={ shadowRootHost }
             >
                 {renderDropdownContainer()}
             </FocusLock>